Move weaver files into experiment folder
